Fix back-end topic styling and HTML block matcher
This commit completes the parser consolidation refactor by replacing
front-end topic name styling with back-end tokenization. Now, the
display topic name is sent in the topic JSON along with tokens which
are rendered as part of the h1. In addition, to avoid rerenders, now
header elements are created for every topic that is being displayed as
root, and these headers stick around the DOM and are hidden and shown
as part of the displayPath code.

HTML elements and blocks are now consolidated into one matcher and token
type. This is made possible by doing light HTML validation to determine
if a block spans multiple lines. Previously an HTML block was a series
of lines that began with HTML tags, but now we look at each potential
block and check if the outer tag type is balanced, or if it is one
single tag like a single or multi-line <img> tag, and if so we make one
HTML token.