Bring back external link icon and handle break line behavior
The icon makes it clearer which links go to an external site.

The problem with breaklines is that we never want the last word of the
link text to be separated from the external link icon, eg:

Word1 Word2[Icon].

We don't want the spacing to fall out that there is a newline after
Word2 before the icon, or after the icon before the period.

So, on page resize or page render, we check to see if every external
link icon that is visible is vertically aligned with the text that came
before it, and the text that comes after it, if that text is connected
punctuation. If there is misalignment, that means that a line break has
been inserted after the last word or after the icon, and so to prevent
this, we insert a break tag before the last word, so that it falls out
together on the next line with the icon and punctuation. Then,
on any page change, we remove all of these break tags and recalculate
which ones are necessary.

This commit also introduces set fonts for the build process, which was
necessary in order to standardize line rendering in order to test the
external icon break line behavior.