Implement better solution to external link icon line breaking
When an external link icon svg is added after the text of a link,
sometimes the last word of the text and the svg get broken onto
separate lines, or, the svg gets separated from closing punctuation like
"." after the icon. Previously we checked the relative heights of the
elements and inserted a break tag before the last word if any of these
elements were falling out on their own line. Now however, we add padding
to the link tokens before the svg, and then a negative margin to the
svg, so that the padding extends beyond the svg and terminal
punctuation, and if the padding doesn't fit on the line, the entire
last word and svg and terminal punctuation will get wrapped as one unit.

(Based on this SO answer: https://stackoverflow.com/a/25857961/5899236)