Black link hovers and external link icon underlining
First, hover should make link underlines black.

Second, this commit introduces a lot of subtle handling of external
link icons. When the link is selected, we want the blue background to
extend beyond the icon entirely, and, we want the icon to be underlined,
but, we want the underlining to extend only to the end of the base of
the icon, whereas the background box should include even the extending
top of the icon. So, we need to add space characters until the end of
the base of the icon so that these can take a text decoration property,
then, we want the a-tag to include padding at the end that extends the
highlighted background but does not extend the underlining.