Sibling selection should trigger scroll
Previously it seems that going from a link to a sibling link in the
same paragraph would not trigger the scrolling behavior. It seems that
if there is no delay between the page render and the scroll, then the
page content might be loaded but not rendered ie visible, and so there
isn't any height to the page to scroll to, but if there is a delay then
the page renders and has height.

This commit also fixes a bug wherein when loading link selections from
the history or session store, we would verify the metadata pointed at
a real link by trying to find a link by that description before page
load, but no links may be in the DOM at that point. So, we should
instantiate a link object from the metadata whether the metadata is
valid or not (eg, even if the content has changed so that the old
link selection no longer exists,) and then when we try to select the
link, if no such element exists this will be caught the same way that
selectALink callback instantiated links have to be caught in the event
that on render there is no such link, which is handled in a try-catch
block at the top of displayPath. Now, two parts of displayPath
redundantly assign the link selection in the event of a null link
selection at a multi-paragraph path, first L15 where we do it
explicitly, and then L34 where we select the lowest parent link as we
climb the tree rendering the paragraphs of the path. The added L15 is
just to make it clearer what the logic is regarding how we select the
link to a reader of the function.