Rename link.paragraphPathWhenSelected to link.path
Previously there were multiple path concepts related to links, eg the
path to the paragraph containing the link, or the path to the target of
the link (if it is local, global, or import), and thirdly the path of
paragraphs to display when the given link is selected. When previously
a selected import reference previewed the target path but used the URL
of the enclosing paragraph, these three concepts could vary, but now
they always coincide so we can simplify the method calls.