Have bulk --sync initiate build on start with abridged message
Previously it was unclear if running canopy bulk --sync should initiate
a build, because perhaps the user is expressing that they intend to fix
invalid data but recognize that it is not valid now and don't want to be
bothered by existing error messages. However, it is also confusing to
initiate a `watch` process and not have the build reflect the most
recent data. Also, starting a sync process and getting the full build
log seemed to distract from the action that was being done, starting the
watch process rather than doing a build.

So, now we build on start, but we do it without the regular logging and
instead log messages that are specific to the situation of the initial
build, explaining to the user why a build is occurring when they might
not understand why it is happening, and showing any errors with this
context.