Fix links in hashUrl mode and expand browser tests
Previously path instances didn't care about the global hashUrls option,
and stringified the path the same way whether it was set or not, and
then when setting the URL, added a '#' or not depending on context,
but this meant that opening a new tab from a URL opened to a non-hashed
URL, so now, we add the hashes at the instance-stringification stage
rather than at the URL-setting stage.

This commit also expands the browser tests that are done on hash URLs
in two ways - first, for a build with hash URLs running on a  node
server, we now check whether opening a new tab works, which tests
whether the link as prefaced with the hash properly.

Second, and more importantly, we test that if someone is running a site
with hashUrls on a static assets server, we test that opening new tabs
continues to append hashes so that requests hit the root directory and
get served the `index.html` file, rather than requesting a non-existent
folder path, which is the concern that motivates using hashUrls in the
first place. So long as the user is using the Canopy express server,
even if the new tab fails to retain the hash character, the system will
still serve the original `index.html` file properly, so it isn't such
a big deal.

Also, the project path prefix is incorporated into Path#string, and
is included in several places where paths are constructed outside of the
path class, eg link hrefs and JSON data paths, and browser tests are
added to test project path prefix and hash URLs with static servers.