Allow advanced link syntax
Often one wants to make a link where the key and the display text are
related or only slightly different, and having to write the link text
twice seems redundant, eg see [[the answer to the question|the answer]],
so now, there are two new syntaxes for expressing this.

First, you can use a display substring using double pipe syntax. Eg,

[[the answer to the question|the answer]]

would be

[[|the answer| to the question]] - this causes the display text to be
"the answer", whereas the link is to the topic "the answer to the
question".

Secondly, you can interpolate keys with slight differences eg:

I want to go to [[the |US ||treasury]] - this would produce a link
whose text is "the treasury", but whose topic reference is to the topic
"the US treasury". The rule for interpolation is that the link contents
is divided by the pipes, and the components are interpreted as:

shared|key|display|shared|key|display etc,

so [[A|B|C|D|E|F]] would have a topic reference of "ABDE" and a display
text of "ACDF".