Fix finicky link selection persistence spec
Previously, the playwright spec "Last link selections are preferred when
going down" was failing occasionally. This was because the Link#select
method added the selected link class to the new link before it stored
the meta data of the selected link, and because the feature spec was
waiting on the class change, in some cases it would continue interacting
with the page after the class had changed before the metadata was
stored.

In order to avoid this problem in future, feature specs should assert
that the selected paragraph has changed, rather than the link, because
this now occurs last in the displayPath function. Additionally, in
Link#select now the class is added after the metadata has been stored,
so even if you assert on the link selection class it should still wait
until metadata has persisted (although it would not be waiting until
the actual page had rerendered, which is still likely to cause
unexpected behavior.)