Fix several problems with scroll behavior
Multiple problems were previously affecting the automatic scroll
behavior.

First, modern browsers have an "overflow-anchor" feature which detects
changes to the off-screen content and scrolls the page automatically to
keep the viewport constant, however, in Canopy usually content cannot
change above the viewport, and the browser's automatic scrolling was
preventing the call from displayPath to window.scrollTo from working,
so, we add the "overflow-anchor: none" property to the Canopy container
to disable this behavior.

Secondly, we were previously adding padding to the bottom of the page so
that removal of content below the viewport never caused the bottom of
the body to come into view, which would cancel an ongoing scroll, but,
this padding was previously applied to the selected section element,
which became invisible until the next section was selected, and so
there might fail to be padding for long enough for a scroll to be
canceled, until the selected section class was added to another
section, so now, we add the padding to the Canopy container itself
which never becomes invisible.

Thirdly, we had been using window.screen.height to measure the screen in
order to position the selected link at eye-level, but this measure
seems to be counting the number of hardware pixels on the screen, not
the variable number of content pixels, and so when the browser is
zoomed, window.screen.height does not change, but window.innerHeight
does, so now we use that.

Fourthly, the issue of losing padding below the last section element was
made possible due to the recent finicky spec fix - previously we had
been waiting on link selection to change in order to wait for the page
to render, but the selected link class changes before the page is fully
rendered, so the section selection was moved last in order to have a
reliable cue as to when the page is finished rendering, but an 
unexpected consequence of this was that the scroll was for the first
time being initiated before section selection, and because page-bottom-
padding was added based on the selected section class, there was now a
period where the class was not applied to a visible element, and so the
page momentarily lacked bottom padding, which canceled any ongoing
scroll. All this was only possible because the class was now applied
after the scroll was initiated, so now, we move the scroll last just
to be safe, because it is fine for playwright to continue a spec once
the class is applied, before the scroll initiates, and it is more
predictable for all page rendering to finish before the scroll is
initiated.

It seemed as though adding a timeout might be necessary to start the
scroll only after the new page had rendered, but given these three
changes it seems all outstanding issues with scroll are addressed,
but that would also have worked to prevent some of these issues.