Log build errors to HTML and consolidate build code
Previously, `watch` had an option not to build initially, and bulk
sync used this option to create custom logging for the initial build
which the user did not directly initiate. However, this logging
difference was general not just to sync mode, but to all cases of watch,
so it was incorporated into the watch code.

Additionally, build now has two levels of wrappers around the build
script, first, errors are caught and written to the index.html file if
we are in `sync` mode, and then the error is rethrown so that the error
can be caught by the given caller, either the initial build or all
subsequent builds, and should get logging specific to how it is being
called.