Change additional link syntax to curly braces
This commit introduces a further change to the additional link syntaxes.

Previously, links used pipe interpolation syntax ie
[[shared|target|display|shared]], but this required the user to remember
the order of these and sometimes resulted in cryptic empty pipes eg
[[||x|y]].

Instead, now there will be a more expressive syntax that should be
easier for people to remember.

In the new syntax, {X} represents text to be added to the display only,
and {|X|} represents text to have the display exclusively set to.
(Multiple instances of exclusive syntax will be concatenated.)

Conversely, {{X}} represents the text to add to the target, and {{|X|}}
represents the text to set the target to exclusively.

Thus single braces {} represents display text, double braces {{}}
represent target text, and |pipes| represent exclusivity.

Finally, there is a new interpolation syntax, such that {A|B} causes
"A" to get added to the target and "B" to be added to the display.

Hopefully these options will give a more visually semantic way of
expressing the ideal target/display text.