Clarify multi-line subtopics in README
Previously, all examples used inline syntax, which can be unclear for
multi-line paragraphs. This new version makes it clear that multi-line
paragraphs can optionally have the topic/subtopic key on its own line,
while the single-line examples make it clear that inline is the default.
Thanks to @agibralter for the suggestion!