Modify scrolling behavior to accommodate unloaded images
Previously, if there were images on the screen, displayPath would scroll
to the appropriate link, and then the image load would jump the scroll
position to be at the incorrect point.

We have to accommodate two major cases:

- It is the initial page load and we are doing an 'auto' scroll to the
  selected link. When images load, that scroll should also be 'auto'

- The user has made a link selection, and the previewed paragraph
  contains an unloaded image. If we initiate a scroll at first, then
  when images load we may still be in the middle of that scroll, and we
  should not initiate a new scroll during that scroll. So the post-load
  scroll should not interrupt the initial scroll, and should also be
  'smooth'.

To accommodate the second case, we could try to detect when the pre-load
scroll has stopped before initiating a second scroll. However, it is
easier instead to have displayPath detect unloaded images and in that
case never initiate a scroll in the first place, and later have the
load event on the last-loaded image trigger the scroll.

In order to accommodate the first case, we now mark in the displayPath
code which scroll should be used next when images load.

Both cases should cause a scrollStyle to be stored on the
canopyContainer element.

Previously we checked for image.complete and image.naturalHeight > 0,
in order to avoid waiting on scrolling for images that will not
ultimately affect scroll, but all new images have naturalHeight == 0
until the asset is received, and so this caused images that would later
break the scroll to seem unproblematic, and so displayPath didn't save
a scrollStyle for a deferred scroll in those cases, and so when the
image ultimately did trigger a scroll, it defaulted to 'auto', violating
the requirements for case 2.