Change scrolling behavior and preload images
In a recent commit the issue was addressed that sometimes images load
slowly and when they are inserted into the DOM they mess up the scroll
of the page. At that time, the solution used was to delay the initial
link-selection page scroll until the images loaded so that there
wouldn't be two jumps, and so there wouldn't be a slow scroll that got
interrupted mid-way. However, this approach didn't work for very slowly
loading assets, because the user is left for seconds without a link
selection.

So, this commit introduces a few changes to address this problem. First,
now we go back to scrolling even before the images are loaded. In order
to avoid a double-jump, now we keep track of what scroll behavior was
most recently used, eg 'auto' for an initial page load, or 'smooth' for
a user-initiated page load. That way, if the scroll is
'auto', we can do another 'auto' scroll on image load, and if it is
'smooth', we can do another smooth, and so even if the page is still in
 the middle of a smooth scroll, if an image loads and knocks the
 viewport up or down, an on-load event handler will reinitiate a smooth
 scroll so that there isn't a hard stop to the ongoing smooth scroll
 even if the page contents changes during it.

Additionally, we now put a 700px height on images until they load, which
helps avoid a visual jump because having the screen go from white to
image isn't so disorienting, whereas going from text to image is an
obvious viewport shift. 700px is a good number to pick because the
selected link is either in the paragraph of the image below the image,
or it is above the image, either in the same paragraph or the parent
link. So, if the selected link is below, and given the selected link is
scrolled to around 25-30% of the page, 700px guarantees for most
monitors that this will be more than the visible component of the image
on the screen above the selected link, regardless of how large the
image ends up being, because the number of visible pixels above the
link is a constant function of screen size.

Conversely, if the selected link is above the image, firstly, a late
load doesn't disrupt the position of the selected link even if the
content of the page below the link jumps on image load, and secondly,
700px should be enough that on almost all monitors, the screen below
the selected image will be blank until image load, avoiding the display
of text that gets replaced with image on image load.

Lastly, now when JSON is eager loaded at page load / global link
render time, we scan that JSON for image URLs and preload them which
hopefully will sometimes avoid the image-load page-jump in the first
place.