Bug: Subtopic subsumption check permits cycles
We have a check in the parserContext to make sure that a given subtopic
is subsumed by the root topic. It is used in a few places, eg to warn
regarding orphan subtopics, error on import references to unsubsumed
subtopics, and to find redundant subsumed local references.

Previously, this check did a depth first search permitting cycles,
because later there is a check to prevent cycles, ie the check for
redundant local references. However, until this commit it was possible
to hit a stack overflow where this check gets caught in the cycle before
the redundant local references check identifies it and errors.