Bug: Validate that topic is direct child of previous subtopic
Given the path 'A/C', we should find a section element C that is a
direct child of section element 'A'.

However, in the case where element A had a subtopic B that also had a
global reference to B, it was possible for the path 'A/C' to
non-deterministically return the element 'A#B/C', because the DOM
query previously used was requesting an element "C" that is a descendant
of A, not specifying that C must be a direct child of A, thus creating
an ambiguity when two subtopics reference the same global reference.

Similarly, if the current path would be 'A/C#D', we would be looking for
a D that is a descendant of a C that was a direct child of A, and not of
a subtopic of A.

This change uses the `:scope` keyword to require that there be a direct
child topic for the subtopic of the next path segment, and it therefore
expands the query to assert the existence of such a topic that has a
descendant of such a subtopic.

This commit also consolidates the root element case into the same
iterative process by considering the canopyContainer to have a path
depth of -1.