Improve table handling
This commit introduces several changes to table handling behavior.

First, we fix two bugs in the table matching code. Previously a table
that was not at the end of the paragraph would have a trialing newline
which would get included in the matched table string and then when we
split on newlines to get the rows of the table, this trailing newline
produced an imaginary empty last row to the table which seemed to have
the wrong number of elements and thus invalidate the table match, so now
we remove the trailing newline after verifying it exists; (the presence
of the newline character or end of string is mandatory for the table
match, so we want to verify it is there, but not include it in the match
string we are converting into a table token.)

The second bug is that previously we were rejecting lines with cells
what were all delimiters with the regex /^\s*[=#-|]+\s*$/, but the
hyphen wasn't escaped so in effect we were rejecting lines that had
characters that were from # to |, ie ASCII 035 to 124, which happens to
include the entire alphabet, 065 to 122, and numbers 048 057 so tables
with alphanumeric cell values were being rejected as delimiter rows.

Thirdly, now tables take up 90% of the width of the page if there is
enough content, and cells will grow to fit content up until that point,
and beyond that point text will wrap.