Allow [:;?] to follow stylized text
We need to distinguish between characters that can follow style
characters, and which cannot.

For example, the following should be valid italic text:

"He wants us to come _now_"

Whereas the following should not be:

"The filename is 2012_beach_photos."

This commit adds ? : and ; to the list of characters that may follow
a style character without causing it to be parsed as plaintext.