Add forward slash to bulk file name in gitignore
As part of the initialization process we create a `.gitignore` file for
the project. If the user runs the `canopy bulk --sync` command, by
default a file will be created with the name of the default topic, and
so we ignore that file by default. However, adding "X" to the gitignore
file apparently ignores all instances of X at any point in the path,
similar to `**/X/**`, which causes the default topic's category
directory to be ignored in the `topics` directory, so we limit this
ignore rule to only apply to the root directory.

https://git-scm.com/docs/gitignore

> If there is a separator at the beginning or middle (or both) of the
  pattern, then the pattern is relative to the directory level of the
  particular .gitignore file itself. Otherwise the pattern may also
  match at any level below the .gitignore level.