Browser back keypress should not prevent default
In general when a user presses an arrow key or the tab key, we want to
prevent default because these keys have special meaning in Canopy and
if we allow the default scrolling behaviors to occur, they may conflict
with Canopy's own internal scrolling behavior. However, one exception
is when the user uses the command/control + left/right key shortcut,
wherein they are obviously intending the default behavior to occur, so
now we handle this as an explicit exception to the prevent default rule.