Allow logging/writing of non-build related sync-mode errors
Previously, the logic by which an error could get caught, logged, and
written to the index.html file for the writer to see during bulk sync
mode editing was specific to errors originating in the build process,
ie parsing the expl files, but not errors originating in the sync
code, eg errors specific to parsing the bulk file. Now, there is a
generic helper function for running a callback, catching any errors,
logging them, and if we are in sync mode, writing the error to the
index.html file.

This change revealed a bug wherein we were erasing the selection file
from disk as soon as it was read, which caused it to be absent in the
case where bulk file syntax errors result in a "rollback" to the
previous bulk file state. In such a case, the old bulk file and the old
selection file should remain on disk as they were before the problematic
bulk file change. So now, we only delete the old selection file when
we have made it successfully to the point where we are deciding whether
to write a new selection file.

This commit also fixes the code that makes sure on bulk-file-save that
we only log about the change if there really was a change, so now on
bulk file creation we keep track of the old bulk file string, and check
if the new one is different.