Increase table font size and list padding
This commit increases the font size for HTML tables which was being
overridden by the default HTML table style having greater specificity
than the `#_canopy` container font size declaration.

Additionally, now LIs only get padding on the bottom, except for the
last one, and the list as a whole gets margin so that the space above
the list matches the space below.