Require explicit import syntax in case of multiple global candidates
Canopy users who create import references must enable the parser to
identify the global reference to which each import reference belongs.
They can do this by writing an import reference using explicit syntax,
for example [[United States#New York]], or, they can declare a global
reference and an import reference in the same paragraph, and rely on
implication for the parser to identify the relationship.

Additionally, once all the import references have been created, they are
validated to ensure that every import reference has a matching global
reference in the same paragraph. This is necessary in case explicit
import references are declared without matching global references.
(Implicit declarations would only have worked if this check succeeded
initially.)

However, these two operations, implicit import global assignment and
validation of global import matching did not work in the case where
an import reference and the matching global reference were in different
tokens, eg different cells of a table.

Now, when looking for an implied import reference's matching global
reference, we look in the entire paragraph text, and not just the
currently parsing token eg a list or table. Additionally, when checking
that import references have matching global references, we are using
a record of all the global and import references from the current
paragraph, rather than just the tokens of the currently parsing token.

To simplify this change, and to avoid the potential for misattributing
import references to the wrong global reference, we are deprecating a
prior functionality wherein an implicit import reference in a paragraph
with two candidate global references would be assigned to one based on
proximity. Now, we continue to use context to differentiate between
local references and import references, eg noticing which
interpretations produce redundant local references and avoiding them,
but no longer do we guess which global link in a paragraph is the owner
of an implied import reference, which simplifies the aforementioned
changes.

At a later time, the code may be simplified by requiring all import
references that are ambiguous, eg, import references that match
subtopics of the current topics, to be declared explicitly.