Add error message for ambiguous local/import references
Previously there was one error message for both the case where you have
two redundant local references neither of which could be anything
except a local reference, and the case where one or either could be a
global or import reference. Now, there is a new error message that
identifies specifically which of the matching references could be which
other types of link, and offers specific suggestions for how to address
the issue.