Bug: error message character count off for multi-line tokens
Previously when errors occurred inside table cells or list items,
we were calculating the current position by looking at where the table
or list started, and then counting the number of characters within the
current LI/TD, without taking into account previous lines or cells on
the current line.

Now, it is the responsibility of the function calling parseText
recursively to specify at what character offset we are starting that
parse, and to specify what significance to assign newlines, as in the
case of block quotes where a newline implies a leading `> `.

`parse_paragraph_style_examples.test.js` is deprecated because it relied
on mocking parserContext which required keeping the mock in sync with
the class, which was a source of error.