Allow style characters within words
Previously, *bold* worked, and *bold words* worked, but *b*old did not,
in line with most markup character implementations. However, now we are
going to assume that matching style characters within a word are
intentional, and require the user to escape eg
`filenames\_with\_underscores.`

However, style characters within a word can only match style characters
within that same word, but not eg `a*b c*d`, because the likelihood
of unintentional matches is much higher.