Replace category-note-ignoring with "?"-escaping
When using bulk mode, a user might make a category notes file that
begins with a note that begins with a question. For example:

```
[A/B/C]

This is a note that begins with a question? And here is more.
```

Initially, this would recognize as a topic and would mistakenly generate
a JSON file with the question name.

Previously, to deal with this problem, the parser would ignore
'category notes' files, ie files that matched the name of their
most proximate enclosing category, if the topic name did not match the
file name. However, this introduced bugs for example where a topic name
had markdown and the filename did not, yet still should match.

Therefore, a new method is introduced to deal with the initial problem.
Now, all files with first paragraphs with valid keys become topics.
However, when a user is using bulk mode and saves a note with an
initial question, the question mark gets escaped when it is saved on
disk. This is preferable, because it doesn't require any changes to the
parser, only the bulk mode code, and we are preserving the information
of it being intended as a note while we still have that information
from the bulk file (where question-initial-notes and topics look
different), before it is destroyed upon saving to disk
(where question-initial-notes and topics look the same.)

Additionally, this commit introduces Topic.for as an alias to class
initialization for when the user just wants the class for one-time use.

Finally, we remove tests relating to the old behavior and add new ones
for the new approach.