Trim topics caps form
Now, when a topic is put in "caps" form to compare it to another topic
for reference matching purposes, we ignore white space before and after
the text. This is because the user might have written a link like:

[[{|the} United States Treasury]]

And we want to allow a space before the first real word for readability,
even if technically this produces a link target of " United States
Treasury" with a leading space.