Add fixed table layout
We want tables to adapt to their content such that tables without
much content do not span the full width of the page. However, we also
don't want individual cells to adapt to extremely to their content,
in the case when one cell has much more content than the other. This
combination of properties seems to accomplish this, making the overall
table sensitive to content-size, but the individual cells similarly
sized to one another.