Scroll to center previewed paragraph
Previously the scroll algorithm was to position the selected link around
25% of the viewport so that the previewed paragraph would also be
visible. However, in cases where the paragraph enclosing the selected
link was very large, the previewed paragraph might not be fully visible.
Additionally, this made it more confusing for beginners to understand
that the paragraph they should be focused on was the previewed
paragraph, not the parent paragraph.