Prefix files from bulk mode with order numbers
Previously, the topics of a category were unordered, and if one created
several files in bulk mode and then reloaded the bulk file, the topics
would reorder themselves alphabetically. This was problematic in the
case of e.g. topics that are sequential lines of a text or entities in
an ordered series. If the user wanted to preserve the order of the
topics in the bulk file across sessions, they would have to make
a set of numbered categories of one topic each.

Now, we take advantage of the fact that topic file name is not used for
anything to prepend a padded integer before the topic name that
indicates the topic's order in the bulk file. When the bulk file is
generated, filenames are ignored except as the basis of the initial
alphabetization process that precedes bulk file generation, after which
topic keys within the file contents are used to generate the file etc.

In a later commit, we may use the convention of numbered topics to
replace the concept of category notes with named note files which can be
named after the initial note text and can be guaranteed to be unique due
to file numbering.