Make ordered categories opt-in and detect on bulk file generation
Previously every category created in bulk-file mode was considered
ordered and would have numbers prepended to retain their order from the
bulk file. However, this seemed excessive as only in rare cases does
topic order matter within a category. So, this commit introduces the
ability for a writer to request that Canopy keep track of a certain
category's topics, and the ability for Canopy to detect on bulk file
generation that the topics of a category were previously ordered.

Now, if a user would like to indicate that the topics of a category are
ordered, the following category syntax can be used in the bulk file:

```
[A/B/C/#]

* Topic A: XYZ

* Topic B: XYZ
```

This will produce the files `A/B/C/1-Topic_A.expl` and
`A/B/C/1-Topic_B.expl`

```