Remove ordered categories but keep category note removal
The last two commits introduced the idea of ordered categories, and then
used the resulting uniqueness of filenames to replace category note
files with individual note files, and here we are preserving that latter
change by adding individuating suffixes in the case that two note files
get the same filename.

Ordered categories were removed because the same problem that occurs
with topics would occur with the categories themselves, and there is no
easy way of adding ordering information to category names without a
visual change to the bulk editor.