Preload JSON instead of using data attribute
Previously we were adding the default topic's JSON to a data attribute
so that it was available when the JS loaded, but putting it in the
HTML delayed the CSS and JS requests, so now we still reload it from
the HTML but parallelize the request with the CSS and JS and access the
asset from the existing JSON fetch code.

In order to get the preload to work, the server code had to be edited
to use cors headers so that the fetch recognized the preload was for
the same requested data.