Image token source URLs should have project path prefix
Previously, only path URLs generated in the app automatically were
prefixed with project path prefix, but if the user specifies an asset
that is hosted locally, we would want to accommodate that automatically
so that the user doesn't have to couple information about their service
provider to the data of their project. Rather, we would like a relative
path to a local asset to work regardless of whether project-path-prefix
is specified at build-time.