Changes to fix JSON preloading
We preload the default topic's JSON in the initial index.html but it
needs to have the same request headers for the fetch to hit the cache.

This commit also separates custom.html into head.html and body.html so
that the user can specify where they want their HTML to run.