Reset original styles on HTML image tags
If the user declares a raw HTML image, we hide it until it is loaded
and give it a fixed height so that when it renders it doesn't cause the
page to jump. However, if the element had inline height or width
attributes, those weren't preserved so now we keep track and set them
back to their original values.